+++ contacts/src/rdflib/ContactsModuleRdfLib.ts
@@ -8,6 +8,7 @@ import {
   CreateNewGroupCommand,
   FullContact,
   FullGroup,
+  NewContact,
   RemoveContactFromGroupCommand,
 } from "..";
 import { AddressBookQuery, ContactQuery } from "./queries";
@@ -50,6 +51,8 @@ export class ContactsModuleRdfLib implements ContactsModule {
       this.fetchNode(groupIndex),
     ]);
 
+    await this.fetchAll([nameEmailIndex, groupIndex]);
+
     const contacts = query.queryContacts();
     const groups = query.queryGroups();
     return {
@@ -64,13 +67,51 @@ export class ContactsModuleRdfLib implements ContactsModule {
     return fetchNode(this.fetcher, node);
   }
 
+  private async fetchAll(nodes: (Node | null)[]) {
+    return Promise.all(nodes.map((it) => this.fetchNode(it)));
+  }
+
   async createAddressBook({ containerUri, name }: CreateAddressBookCommand) {
     const operation = createAddressBook(containerUri, name);
     await executeUpdate(this.fetcher, this.updater, operation);
     return operation.uri;
   }
 
-  async createNewContact({ addressBookUri, contact }: CreateNewContactCommand) {
+  async createNewContact({
+    addressBookUri,
+    contact,
+    groupUris,
+  }: CreateNewContactCommand) {
+    const contactQuery = await this.executeCreateNewContact(
+      addressBookUri,
+      contact,
+    );
+    await this.executeAddContactToGroups(groupUris, contactQuery);
+
+    return contactQuery.contactNode.uri;
+  }
+
+  private async executeAddContactToGroups(
+    groupUris: string[] | undefined,
+    contactQuery: ContactQuery,
+  ) {
+    const groupNodes = (groupUris ?? []).map((it) => sym(it));
+    await this.fetchAll(groupNodes);
+
+    const groupUpdates = groupNodes.map((groupNode) => {
+      const groupQuery = new GroupQuery(this.store, groupNode);
+
+      const operation = addContactToGroup(contactQuery, groupQuery);
+      return executeUpdate(this.fetcher, this.updater, operation);
+    });
+
+    await Promise.allSettled(groupUpdates);
+  }
+
+  private async executeCreateNewContact(
+    addressBookUri: string,
+    contact: NewContact,
+  ) {
     const addressBookNode = sym(addressBookUri);
     await this.fetchNode(addressBookNode);
     const operation = createNewContact(
@@ -78,7 +119,8 @@ export class ContactsModuleRdfLib implements ContactsModule {
       contact,
     );
     await executeUpdate(this.fetcher, this.updater, operation);
-    return operation.uri;
+
+    return new ContactQuery(this.store, sym(operation.uri));
   }
 
   async readContact(uri: string): Promise<FullContact> {